Adding types back for overflow protection and clarity
Auto type deduction wasn't working properly for the percent calculation
in Crypto::executeCipher. The problem was that the percent variable type
was being deduced without a cast before the later calculation with
larger types, causing overflow. Having the types here is more clear than
adding in casts and continuing to use auto.