Switched Botan to amalgamation build
Switched Botan to amalgamation build because it's a lot easier for
another developer to build Botan without having to worry about binary
compatibility with the compiler I built the libraries with that were
previously in the repository. There are still multiple versions of the
amalgamation build, since Botan decides what features and optimizations
it can use by platform. Making this switch solves most of the
configuration problems associated with having a separate header include
set and libraries. Now there are also tests in the pro file to determine
what operating system is currently being built, and macro checks in the
Crypto class where a platform-specific version of Botan is included.