Added Android support
Kryvos could already build for Android, but like many desktop
application ports to mobile, it didn't look very good. This is an
attempt at fixing that. Now there is a new MainWindow class that
contains the common features used in a desktop application and mobile
application. The new classes DesktopMainWindow and AndroidMainWindow now
inherit from it, and perform their respective platform-specific
modifications to the GUI. The main thing that looked bad on Android was
button sizes and a few borders that don't make sense on a mobile device.
Most of the colors and font details are specified in Qt stylesheets, so
a new stylesheet was added to address this problem. The font-size of
text labels and buttons was adjusted to resize the app to a more usable
state on Android. Other widget classes were updated to add access
functionality for resizing purposes. I originally was going to create an
inheritance hierarchy for them, but decided the changes were applicable
to any custom platform-specific size adjustments. Finally, touch
scrolling was added to the file list and message text edits. I disabled
overshoot because overshoot makes scrolling on small widgets difficult.