Added support for Mac OSX and iOS
Generated and added Botan amalgamation builds for Mac OSX and iOS.
Changed source files to support proper usage of templated % with
QStringBuilder that was caught by clang compiler. Switched
AndroidMainWindow to TouchMainWindow, as it currently contains the same
changes for Android and iOS. Added Mac icns file and fixed paths for
stylesheets.